backend/requirements.txt
asgiref==3.8.1
django==4.2.21
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
gunicorn==20.1.0
django-environ==0.12.0
psycopg2-binary